🚚 컴포넌트 파일 폴더 분리
