+++ src/app/components/common/ClientContent.tsx
